requirements.txt
websockets>=12.0
psycopg2-binary>=2.9.9
requests>=2.31.0
cryptography>=41.0.0

